refactor(test): change component signature
